add typer styling
